requirements.txt
# 核心依赖
flask==2.0.1  # Web框架
flask-sqlalchemy==2.5.1  # ORM数据库支持
flask-migrate==3.1.0  # 数据库迁移工具
werkzeug==2.0.1  # Flask依赖，兼容版本

# 安全和认证
flask-login==0.5.0  # 用户认证
flask-wtf==1.0.0  # 表单验证
flask-cors==3.0.10  # 跨域资源共享
flask-limiter==2.0.3  # 请求限制
flask-jwt-extended==4.3.1  # JWT扩展支持
email-validator==1.1.3  # 邮箱验证
bcrypt==3.2.0  # 密码加密
cryptography==36.0.1  # 加密库
pyjwt==2.3.0  # JWT令牌支持
python-dotenv==0.19.2  # 环境变量处理

# 数据库驱动
pymysql==1.0.2  # MySQL驱动
psycopg2-binary==2.9.3  # PostgreSQL驱动，可选

# UI和前端
flask-bootstrap==3.3.7.1  # Bootstrap支持
flask-moment==1.0.2  # 时间本地化
flask-mail==0.9.1  # 邮件支持

# 文档和文件处理
Pillow==9.0.1  # 图像处理
reportlab==3.6.8  # PDF生成
PyPDF2==1.26.0  # PDF处理
python-docx==0.8.11  # Word文档处理
openpyxl==3.0.9  # Excel文档处理
python-pptx==0.6.21  # PowerPoint处理
pdfrw==0.4.3  # PDF读写
xlrd==2.0.1  # Excel读取
xlwt==1.3.0  # Excel写入

# 系统和工具
psutil==5.9.0  # 系统资源监控
requests==2.27.1  # HTTP客户端
markdown==3.3.6  # Markdown支持
jsonschema==4.4.0  # JSON模式验证
python-dateutil==2.8.2  # 日期处理工具
fonttools==4.33.3  # 字体处理工具
brotli==1.0.9  # Brotli压缩，用于字体子集化

# 中间件和反向代理
gunicorn==20.1.0  # WSGI HTTP服务器

# 开发和测试工具
pytest==6.2.5  # 单元测试
coverage==6.2  # 代码覆盖率测试

# 注意: 如果使用Windows系统，psycopg2-binary可能需要额外的编译环境
# 如果遇到安装问题，可以先注释掉，或者使用纯Python替代方案 